refactor: 패키지를 아키텍처에 맞게 변경한다
