+++ src/main/java/bridge/application/port/service/BridgeGame.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.application.port.service;
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
